Clean up the code by making it more concise
Please be _sure_ to check this _before_ merging, in-case I've missed
something. I checked bits and pieces, and some I assume will be fine, -
as I've done them many times. You never, though.

The general gist, is that the `Do()` function serves to avoid code
repetition, and, as your comment suggested you wanted, it's now using
`find`.

However, to keep the output on the one line, being able to see each
file as its being processed, is done here thanks to a few things:

  * The `setterm` command disables line-wrapping, avoiding new lines.

  * The `-printf` flag's `\r` clears from cursor position to column 0.

  * The `echo` creates the final new line character.

Line wrapping _should_ be re-enabled before `Do()` finishes.

The whole _default password if none is given_ thing is, at least in
this case, better done using `:-`, as shown, which the shell will end
up seeing it as _if not set, then instead expand to this_. No
assignment takes place, by the way. Also, `:+` is the opposite.

Replaced the main `if` statement with a `case` statement, to condense.