Optimize `monitor.sh` (fewer processes)
The process by which this script fetches the RAM usage is now done only
with `free` and the Bourne shell, so no more AWK needed.

It's a shame pipes are needed, since a subshell is created for each one
used, although it looks like that behavior is shell-dependent.

  "The POSIX standard allows executing any of subcommands in the
  current shell, rather than subshells, ..." - yash(1)

Using pipes can wind up to be a lot of new processes. In Bash, though -
using `<<< ""` (herestring) is done within the current shell. This is
part of the reason why, in Bash, I favor this sort of while read loop:

  ```bash
  while read CurLine; do
    printf '%s\n' "$CurLine"
  done <<< "$Data"
  ```

Plus I find it easier to read, since the while part is the first word.

Hope that was interesting and educational, albeit entirely unnecessary
for a commit message.