Tweak syntax to be more appropriate
The `*|''` part was a fail on my part, because the `''` never seems to
work how I'd like it to. Not entering any commands still seems to yield
the same result, however.

The `[fdLpSb]` part is just cleaner and more appropriate. Using `|`
over and over wasn't needed at all, and actually may have been
ever-so-slightly inefficient -- oops.